+++ src/ocean/net/http/message/HttpHeaderParser.d
@@ -170,7 +170,7 @@ class HttpHeaderParser : IHttpHeaderParser
 
     ***************************************************************************/
 
-    public uint object_pool_index;
+    public size_t object_pool_index;
 
     /**************************************************************************
