+++ src/ocean/util/container/pool/StructPool.d
@@ -96,7 +96,7 @@ version ( UnitTest )
 {
     struct Struct
     {
-        uint object_pool_index;
+        size_t object_pool_index;
 
         size_t i;
         char[] s;
